requirements.txt
dj-database-url==0.4.2
Django==1.11.2
djangorestframework==3.6.3
gunicorn==19.7.1
olefile==0.44
Pillow==4.2.0
pkg-resources==0.0.0
pytz==2017.2
whitenoise==3.3.0
